Airflow/requirements.txt
apache-airflow[gcp,apache.spark]==2.1.3

pyspark==3.2.0
pandas
yfinance==0.1.67
pyarrow